export parameters
export dialogue box is added, as well as the event at which prevent program to export file to a empty directory. Even if the checkbox is checked, if the directory lineedit is empty, it will not be outputting figure. Looking to add an indication showing if the outputting directory is recorded or not.